requirements.txt
Django==4.0.6
djangorestframework==3.13.1
requests==2.28.1
Pillow==9.2.0
requests==2.28.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
